requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
pymongo>=4.7.0
psycopg[binary]>=3.1.0
psycopg-pool>=3.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
numpy>=2.4.1